[Feature] Added all handlers

+++ repository/repositories.go
@@ -7,8 +7,8 @@ import (
 )
 
 type FindAllPage struct {
-	Size   uint
-	Offset uint
+	Size   uint64
+	Offset uint64
 }
 
 type FindResult struct {
